+++ run.sh
@@ -4,4 +4,4 @@
 cd $(dirname $0)
 make backend
 . ~/.kiterc
-spark-submit target/scala-2.12/lynxkite-0.1-SNAPSHOT.jar
+spark-submit --driver-java-options=-Xss32M target/scala-2.12/lynxkite-0.1-SNAPSHOT.jar